pyproject.toml
[project]
name = "agents-league-techconnect-2026"
version = "0.1.0"
description = "Zava Smart Assistant - Enterprise Incident Response AI Agent for Agents League TechConnect"
requires-python = ">=3.11"
dependencies = [
    "github-copilot-sdk",
    "pydantic",
    "pyyaml",
    "python-dotenv",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[project.scripts]
zava = "console_app:main"

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
